% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dendrometers.R
\docType{data}
\name{data_dendrometers}
\alias{data_dendrometers}
\title{data_dendrometers}
\format{
A data frame with 718 rows and 6 variables:
\describe{
  \item{site}{Simoncouche}
  \item{species}{sugar maple and black spruce}
  \item{tree}{tree identifier}
  \item{year}{2017, 2020}
  \item{doy}{day Of a Year}
  \item{width}{width expressed in relative terms}
}
}
\source{
Data belongs to Sergio Rossi
}
\usage{
data_dendrometers
}
\description{
A dataset with dendrometer data from sugar maple and black spruce from
Simoncouche site, Canada.
}
\keyword{datasets}
