% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_site.R
\docType{data}
\name{data_site}
\alias{data_site}
\title{data_site}
\format{
A data frame with 79 rows and 2 variables:
\describe{
  \item{Tissue}{XYLEM or PHLOEM}
  \item{Species}{Fagus sylvatica (FASY), Picea abies (PIAB), Quercus pubescens (QUPE)}
  \item{Site}{Panska reka (PAN), Karst (KRAS)}
  \item{Year}{2011, 2017}
  \item{gom_a}{The initial value for the Gompertz parameter a}
  \item{gom_b}{The initial value for the Gompertz parameter b}
  \item{gom_c}{The initial value for the Gompertz parameter c}
  \item{brnn_neurons}{The number of neurons for BRNN method}
  \item{gam_k}{The k parameter value for GAM method}
  \item{gam_sp}{The sp parameter value for GAM method}
}
}
\usage{
data_site
}
\description{
A dataset with model fitting parameters for different regression methods.
}
\keyword{datasets}
