% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_gompertz.R
\name{double_gompertz}
\alias{double_gompertz}
\title{double_gompertz}
\usage{
double_gompertz(x, a1, b1, k1, a2, b2, k2)
}
\arguments{
\item{x}{independent variable}

\item{a1}{parameter a1}

\item{b1}{parameter b1}

\item{k1}{parameter k1}

\item{a2}{parameter a2}

\item{b2}{parameter b2}

\item{k2}{parameter k2}
}
\value{
a nls object with the function fit
}
\description{
A method to fit the double Gompertz function
}
\keyword{internal}
