% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodformation_fit.R
\name{woodformation_fit}
\alias{woodformation_fit}
\title{woodformation_fit}
\usage{
woodformation_fit(
  data_trees,
  data_site,
  search_initial_gom = FALSE,
  fitting_method = c("gompertz", "gam", "brnn"),
  ID_vars = NULL,
  fitted_save = FALSE,
  add_zeros = TRUE,
  post_process = TRUE
)
}
\arguments{
\item{data_trees}{a data frame with ID variables and wood formation data with columns DY and EWM}

\item{data_site}{a data frame with ID variables and initial parameter values for the selected methods}

\item{search_initial_gom}{logical, should the algorithm to search initial Gompertz parameters be applied?}

\item{fitting_method}{vector of one or more methods to be compared: "gompertz", "gam", "brnn"}

\item{ID_vars}{character vector of variables which indicate column names of ID variables}

\item{fitted_save}{logical, should the fitted curves be saved in current working directory?}

\item{add_zeros}{logical, should zero observations at the beginning of growing season be added?}

\item{post_process}{logical, should the post-process algorithm be applied?}
}
\value{
a list with the following elements:
\enumerate{
 \item $fitted - a data frame with fitted wood formation data
 \item $gompertz_grid_search - a data frame with selected initial parameter values
 \item $gompertz_grid_search_errors - a data frame with unsuccessful cases of gompertz grid search
}
}
\description{
Function fits and compares selected methods for modeling xylem and floem formation data.
}
\examples{
\dontrun{
library(woodformation)

# Load data
data(data_site)
data(data_trees)
simulation_1 <- woodformation_fit(data_trees = data_trees,
     data_site = data_site,
     ID_vars = c("Species", "Tissue", "Site", "Year", "Tree"),
     fitting_method = c("gompertz", "gam", "brnn"),
     fitted_save = TRUE,
     search_initial_gom = TRUE,
     add_zeros = TRUE,
     post_process = TRUE)

}
}
